🚚rename: 도메인 별 폴더 분리
