+++ frontend/src/routes/rewardPage/companyInfo/CompanyInfo.jsx
