+++ frontend/src/routes/rewardPage/receipt/Receipt.jsx
