+++ frontend/src/routes/rewardPage/tutorial/ReceiptTutorial.jsx
