+++ frontend/src/routes/rewardPage/video/Video.jsx
